+++ documentation/helpers/ArgTypes/index.tsx
@@ -1,33 +1,55 @@
 import { Tabs } from '@spark-ui/tabs'
 import { ArgTypes as StorybookArgTypes } from '@storybook/blocks'
-import type { FC } from 'react'
+import { type FC, type ReactNode } from 'react'
 
 interface Props<T> {
-  of?: T
-  subcomponents?: Record<string, T>
+  of: T
+  description?: string
+  subcomponents?: Record<string, any> | null
 }
 
-export const ArgTypes = <T extends FC>({ of, subcomponents = {} }: Props<T>) => {
-  const components = {
-    [of?.displayName as string]: of,
-    ...subcomponents,
-  }
+const ComponentDescription = ({ children }: { children: ReactNode }) => {
+  return (
+    <p className="rounded-md bg-primary-container px-lg py-md text-body-2 text-on-primary-container">
+      {children}
+    </p>
+  )
+}
+
+export const ArgTypes = <T extends FC>({ of, description, subcomponents = null }: Props<T>) => {
+  if (!subcomponents) return <StorybookArgTypes of={of} />
+
+  const { displayName: name = 'Root' } = of // "Root" in case the root component is missing a displayName
+  const subComponentsList = Object.entries(subcomponents)
 
   return (
-    <Tabs defaultValue={of?.displayName} orientation="vertical" className="mt-xl sb-unstyled">
+    <Tabs defaultValue={name} orientation="vertical" className="sb-unstyled mt-xl">
       <Tabs.List>
-        {Object.keys(components).map(name => (
-          <Tabs.Trigger key={name} value={name}>
-            {name}
-          </Tabs.Trigger>
-        ))}
+        <Tabs.Trigger key={name} value={name}>
+          {name}
+        </Tabs.Trigger>
+        <>
+          {subComponentsList.map(([name]) => (
+            <Tabs.Trigger key={name} value={name}>
+              {name}
+            </Tabs.Trigger>
+          ))}
+        </>
       </Tabs.List>
 
-      {Object.entries(components).map(([name, component]) => (
-        <Tabs.Content key={name} value={name} className="py-none">
-          <StorybookArgTypes of={component} />
-        </Tabs.Content>
-      ))}
+      <Tabs.Content key={name} value={name} className="py-none">
+        {description && <ComponentDescription>{description}</ComponentDescription>}
+        <StorybookArgTypes of={of} />
+      </Tabs.Content>
+
+      {subComponentsList.map(([name, { of, description }]) => {
+        return (
+          <Tabs.Content key={name} value={name} className="py-none">
+            {description && <ComponentDescription>{description}</ComponentDescription>}
+            <StorybookArgTypes of={of} />
+          </Tabs.Content>
+        )
+      })}
     </Tabs>
   )
 }