+++ packages/components/checkbox/src/CheckboxIndicator.tsx
@@ -10,7 +10,7 @@ export const CheckboxIndicator = forwardRef<HTMLSpanElement, CheckboxIndicatorPr
   (props, ref) => (
     <CheckboxIndicatorPrimitive
       ref={ref}
-      className="text-surface flex h-full w-full items-center justify-center"
+      className="flex h-full w-full items-center justify-center text-surface"
       {...props}
     />
   )