+++ packages/components/popover/src/Popover.stories.tsx
@@ -8,7 +8,7 @@ import { Popover } from '.'
 import { type ContentProps } from './PopoverContent'
 
 const meta: Meta<typeof Popover> = {
-  title: 'Experimental/Popover',
+  title: 'Components/Popover',
   component: Popover,
 }
 
@@ -40,8 +40,10 @@ export const Default: StoryFn = _args => {
         </Popover.Trigger>
         <Popover.Portal>
           <Popover.Content>
-            Popover contents
+            <Popover.Header>Title</Popover.Header>
+            <p>Are you sure you want to have that cookie now ?</p>
             <Popover.Arrow />
+            <Popover.CloseButton aria-label="Close the popover" />
           </Popover.Content>
         </Popover.Portal>
       </Popover>
@@ -54,7 +56,7 @@ export const Controlled: StoryFn = () => {
 
   return (
     <>
-      <div className="gap-lg mb-lg flex">
+      <div className="mb-lg flex gap-lg">
         <Button onClick={() => setOpen(true)}>Open popover</Button>
         <Button onClick={() => setOpen(false)}>Close popover</Button>
       </div>
@@ -65,7 +67,8 @@ export const Controlled: StoryFn = () => {
           </Popover.Anchor>
           <Popover.Portal>
             <Popover.Content onInteractOutside={() => setOpen(false)}>
-              Popover contents
+              <Popover.Header>Title</Popover.Header>
+              <p>Are you sure you want to have that cookie now ?</p>
               <Popover.Arrow />
             </Popover.Content>
           </Popover.Portal>
@@ -79,25 +82,18 @@ export const Anchored: StoryFn = _args => {
   return (
     <ShowcaseContainer className="!justify-between">
       <Popover>
-        <div className="gap-md flex flex-col">
-          <Popover.Trigger asChild>
-            <Button>Trigger 1</Button>
-          </Popover.Trigger>
-          <Popover.Trigger asChild>
-            <Button>Trigger 2</Button>
-          </Popover.Trigger>
-          <Popover.Trigger asChild>
-            <Button>Trigger 3</Button>
-          </Popover.Trigger>
-        </div>
-
         <Popover.Trigger asChild>
-          <Button intent="secondary">Anchor element</Button>
+          <Button>Trigger</Button>
         </Popover.Trigger>
+
+        <Popover.Anchor asChild>
+          <Button intent="secondary">Anchor element</Button>
+        </Popover.Anchor>
 
         <Popover.Portal>
           <Popover.Content>
-            Popover contents
+            <Popover.Header>Title</Popover.Header>
+            <p>Are you sure you want to have that cookie now ?</p>
             <Popover.Arrow />
           </Popover.Content>
         </Popover.Portal>
@@ -114,7 +110,7 @@ export const MatchTriggerWidth: StoryFn = _args => {
           <Button>Check the width of this popover</Button>
         </Popover.Trigger>
         <Popover.Portal>
-          <Popover.Content matchTriggerWidth>
+          <Popover.Content matchTriggerWidth aria-label="Match Trigger Width Example">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
             incididunt ut labore et dolore magna aliqua.
             <Popover.Arrow />
@@ -129,7 +125,7 @@ export const Boundaries: StoryFn = () => {
   const [boundaryContainer, setBoundaryContainer] = useState<HTMLDivElement | null>(null)
 
   return (
-    <ShowcaseContainer ref={setBoundaryContainer} className="!w-sz-256 !h-sz-240">
+    <ShowcaseContainer ref={setBoundaryContainer} className="!h-sz-240 !w-sz-256">
       <Popover>
         <Popover.Trigger asChild>
           <Button>Trigger popover</Button>
@@ -159,8 +155,8 @@ export const Positionning: StoryFn = _args => {
   }
 
   return (
-    <div className="gap-lg flex flex-col">
-      <div className="gap-lg flex">
+    <div className="flex flex-col gap-lg">
+      <div className="flex gap-lg">
         <p className="text-headline-2">Side:</p>
         <RadioGroup value={currentSide} onValueChange={handleChangeSide} orientation="horizontal">
           {['bottom', 'top', 'left', 'right'].map(side => (
@@ -171,7 +167,7 @@ export const Positionning: StoryFn = _args => {
         </RadioGroup>
       </div>
 
-      <div className="gap-lg flex">
+      <div className="flex gap-lg">
         <p className="text-headline-2">Align:</p>
         <RadioGroup value={currentAlign} onValueChange={handleChangeAlign} orientation="horizontal">
           {['start', 'center', 'end'].map(align => (
@@ -189,7 +185,7 @@ export const Positionning: StoryFn = _args => {
           </Popover.Trigger>
           <Popover.Portal>
             <Popover.Content side={currentSide} align={currentAlign}>
-              Popover contents
+              <p>Are you sure you want to have that cookie now ?</p>
               <Popover.Arrow />
             </Popover.Content>
           </Popover.Portal>