+++ packages/components/popover/src/PopoverCloseButton.tsx
@@ -1,0 +1,42 @@
+import * as RadixPopover from '@radix-ui/react-popover'
+import { Icon } from '@spark-ui/icon'
+import { IconButton } from '@spark-ui/icon-button'
+import { Close as CloseSVG } from '@spark-ui/icons/dist/icons/Close'
+import { cx } from 'class-variance-authority'
+import { forwardRef, useLayoutEffect } from 'react'
+
+import { usePopover } from './PopoverContext'
+
+export type CloseButtonProps = RadixPopover.PopoverCloseProps & {
+  'aria-label': string
+}
+
+export const CloseButton = forwardRef<HTMLButtonElement, CloseButtonProps>(
+  ({ 'aria-label': ariaLabel, className, ...rest }, ref) => {
+    const { setHasCloseButton } = usePopover()
+
+    useLayoutEffect(() => {
+      setHasCloseButton(true)
+
+      return () => setHasCloseButton(false)
+    }, [])
+
+    return (
+      <RadixPopover.Close
+        data-spark-component="popover-close-button"
+        ref={ref}
+        className={cx('absolute top-md right-md', className)}
+        asChild
+        {...rest}
+      >
+        <IconButton size="sm" intent="neutral" design="ghost" aria-label={ariaLabel}>
+          <Icon>
+            <CloseSVG />
+          </Icon>
+        </IconButton>
+      </RadixPopover.Close>
+    )
+  }
+)
+
+CloseButton.displayName = 'Popover.CloseButton'