+++ packages/components/button/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.3](https://github.com/adevinta/spark/compare/@spark-ui/button@1.13.2...@spark-ui/button@1.13.3) (2023-06-28)
+
+### Bug Fixes
+
+- **button:** fix hover styles not being applied correctly when using the 'asChild' prop on button ([0be6df0](https://github.com/adevinta/spark/commit/0be6df0031831eadbbf7cab189f97546ef09b030)), closes [#950](https://github.com/adevinta/spark/issues/950)
+
+## [1.13.2](https://github.com/adevinta/spark/compare/@spark-ui/button@1.13.1...@spark-ui/button@1.13.2) (2023-06-23)
+
+**Note:** Version bump only for package @spark-ui/button
+
+## [1.13.1](https://github.com/adevinta/spark/compare/@spark-ui/button@1.13.0...@spark-ui/button@1.13.1) (2023-06-21)
+
+**Note:** Version bump only for package @spark-ui/button
+
 # [1.13.0](https://github.com/adevinta/spark/compare/@spark-ui/button@1.12.2...@spark-ui/button@1.13.0) (2023-06-19)
 
 ### Features